pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idml2docbook"
version = "1.1.3"
description = "Convert IDML files to DocBook."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name="Yann Trividic", email="bonjour@yanntrividic.fr"}
]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "bs4>=0.0.2",
    "lxml>=6.0.0",
    "python-dotenv>=1.1.1",
    "unidecode>=1.4.0",
    "packaging>=25.0",
    "natsort>=8.4.0",
    "pandas>=2.3.2",
    "odfpy>=1.4.1",
]
keywords = ["idml", "docbook", "converter"]

[project.urls]
Homepage = "https://github.com/yanntrividic/idml2docbook"
"Bug Tracker" = "https://github.com/yanntrividic/idml2docbook/-/issues"

[project.scripts]
idml2docbook = "idml2docbook.__main__:main"
idml2docbook-install-dependencies = "idml2docbook.install_dependencies:main"

[tool.setuptools]
packages = ["idml2docbook"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]